fix!: Remove ArgSettings from the API
