+++ src/builder/mod.rs
@@ -21,7 +21,6 @@ pub use action::ArgAction;
 pub use app_settings::{AppFlags, AppSettings};
 pub use arg::Arg;
 pub use arg_group::ArgGroup;
-pub use arg_settings::{ArgFlags, ArgSettings};
 pub use command::Command;
 pub use possible_value::PossibleValue;
 pub use value_hint::ValueHint;
@@ -46,3 +45,4 @@ pub use value_parser::PathBufValueParser;
 pub(crate) use action::CountType;
 pub(crate) use arg::display_arg_val;
 pub(crate) use arg_predicate::ArgPredicate;
+pub(crate) use arg_settings::{ArgFlags, ArgSettings};
