+++ tests/builder/main.rs
@@ -5,7 +5,6 @@ mod app_settings;
 mod arg_aliases;
 mod arg_aliases_short;
 mod arg_matcher_assertions;
-mod arg_settings;
 mod borrowed;
 mod cargo;
 mod command;
